refactor: Lightning interface into own file and LedgerCrud with typing
