+++ cashu/mint/ledger.py
@@ -1,6 +1,6 @@
 import asyncio
 import math
-from typing import Dict, List, Literal, Optional, Set, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple
 
 import bolt11
 from loguru import logger
@@ -19,7 +19,6 @@ from ..core.crypto.keys import derive_pubkey, random_hash
 from ..core.crypto.secp import PublicKey
 from ..core.db import Connection, Database
 from ..core.errors import (
-    InvoiceNotPaidError,
     KeysetError,
     KeysetNotFoundError,
     LightningError,
@@ -32,10 +31,11 @@ from ..core.split import amount_split
 from ..lightning.base import Wallet
 from ..mint.crud import LedgerCrud
 from .conditions import LedgerSpendingConditions
+from .lightning import LedgerLightning
 from .verification import LedgerVerification
 
 
-class Ledger(LedgerVerification, LedgerSpendingConditions):
+class Ledger(LedgerVerification, LedgerSpendingConditions, LedgerLightning):
     locks: Dict[str, asyncio.Lock] = {}  # holds multiprocessing locks
     proofs_pending_lock: asyncio.Lock = (
         asyncio.Lock()
@@ -46,8 +46,8 @@ class Ledger(LedgerVerification, LedgerSpendingConditions):
         db: Database,
         seed: str,
         lightning: Wallet,
+        crud: LedgerCrud,
         derivation_path="",
-        crud=LedgerCrud,
     ):
         self.secrets_used: Set[str] = set()
         self.master_key = seed
@@ -145,122 +145,6 @@ class Ledger(LedgerVerification, LedgerSpendingConditions):
         keyset = self.keysets.keysets[keyset_id] if keyset_id else self.keyset
         assert keyset.public_keys, KeysetError("no public keys for this keyset")
         return {a: p.serialize().hex() for a, p in keyset.public_keys.items()}
-
-    # ------- LIGHTNING -------
-
-    async def _request_lightning_invoice(self, amount: int) -> Tuple[str, str]:
-        """Generate a Lightning invoice using the funding source backend.
-
-        Args:
-            amount (int): Amount of invoice (in Satoshis)
-
-        Raises:
-            Exception: Error with funding source.
-
-        Returns:
-            Tuple[str, str]: Bolt11 invoice and payment id (for lookup)
-        """
-        logger.trace(
-            "_request_lightning_invoice: Requesting Lightning invoice for"
-            f" {amount} satoshis."
-        )
-        error, balance = await self.lightning.status()
-        logger.trace(f"_request_lightning_invoice: Lightning wallet balance: {balance}")
-        if error:
-            raise LightningError(f"Lightning wallet not responding: {error}")
-        (
-            ok,
-            checking_id,
-            payment_request,
-            error_message,
-        ) = await self.lightning.create_invoice(amount, "Cashu deposit")
-        logger.trace(
-            f"_request_lightning_invoice: Lightning invoice: {payment_request}"
-        )
-
-        if not ok:
-            raise LightningError(f"Lightning wallet error: {error_message}")
-        assert payment_request and checking_id, LightningError(
-            "could not fetch invoice from Lightning backend"
-        )
-        return payment_request, checking_id
-
-    async def _check_lightning_invoice(
-        self, *, amount: int, id: str, conn: Optional[Connection] = None
-    ) -> Literal[True]:
-        """Checks with the Lightning backend whether an invoice with `id` was paid.
-
-        Args:
-            amount (int): Amount of the outputs the wallet wants in return (in Satoshis).
-            id (str): Id to look up Lightning invoice by.
-
-        Raises:
-            Exception: Invoice not found.
-            Exception: Tokens for invoice already issued.
-            Exception: Amount larger than invoice amount.
-            Exception: Invoice not paid yet
-            e: Update database and pass through error.
-
-        Returns:
-            bool: True if invoice has been paid, else False
-        """
-        invoice: Union[Invoice, None] = await self.crud.get_lightning_invoice(
-            id=id, db=self.db, conn=conn
-        )
-        if invoice is None:
-            raise LightningError("invoice not found.")
-        if invoice.issued:
-            raise LightningError("tokens already issued for this invoice.")
-        if amount > invoice.amount:
-            raise LightningError(
-                f"requested amount too high: {amount}. Invoice amount: {invoice.amount}"
-            )
-        assert invoice.payment_hash, "invoice has no payment hash."
-        # set this invoice as issued
-        await self.crud.update_lightning_invoice(
-            id=id, issued=True, db=self.db, conn=conn
-        )
-
-        try:
-            status = await self.lightning.get_invoice_status(invoice.payment_hash)
-            if status.paid:
-                return status.paid
-            else:
-                raise InvoiceNotPaidError()
-        except Exception as e:
-            # unset issued
-            await self.crud.update_lightning_invoice(
-                id=id, issued=False, db=self.db, conn=conn
-            )
-            raise e
-
-    async def _pay_lightning_invoice(self, invoice: str, fee_limit_msat: int):
-        """Pays a Lightning invoice via the funding source backend.
-
-        Args:
-            invoice (str): Bolt11 Lightning invoice
-            fee_limit_msat (int): Maximum fee reserve for payment (in Millisatoshi)
-
-        Raises:
-            Exception: Funding source error.
-
-        Returns:
-            Tuple[bool, string, int]: Returns payment status, preimage of invoice, paid fees (in Millisatoshi)
-        """
-        error, balance = await self.lightning.status()
-        if error:
-            raise LightningError(f"Lightning wallet not responding: {error}")
-        (
-            ok,
-            checking_id,
-            fee_msat,
-            preimage,
-            error_message,
-        ) = await self.lightning.pay_invoice(invoice, fee_limit_msat=fee_limit_msat)
-        logger.trace(f"_pay_lightning_invoice: Lightning payment status: {ok}")
-        # make sure that fee is positive
-        fee_msat = abs(fee_msat) if fee_msat else fee_msat
-        return ok, preimage, fee_msat
 
     # ------- ECASH -------
