+++ tests/conftest.py
@@ -14,6 +14,7 @@ from cashu.core.migrations import migrate_databases
 from cashu.core.settings import settings
 from cashu.lightning.fake import FakeWallet
 from cashu.mint import migrations as migrations_mint
+from cashu.mint.crud import LedgerCrud
 from cashu.mint.ledger import Ledger
 
 SERVER_PORT = 3337
@@ -64,6 +65,7 @@ async def ledger():
         seed=settings.mint_private_key,
         derivation_path=settings.mint_derivation_path,
         lightning=FakeWallet(),
+        crud=LedgerCrud(),
     )
     await start_mint_init(ledger)
     yield ledger
